refactor: move settings related files
